refactor: 폴더 계층 구조 정리 및 수정
