+++ src/main/java/gift/entity/UserRole.java
@@ -1,5 +1,15 @@
 package gift.entity;
 
 public enum UserRole {
-    USER_ROLE, MD_ROLE, ADMIN_ROLE
+    ROLE_USER("ROLE_USER"), ROLE_MD("ROLE_MD"), ROLE_ADMIN("ROLE_ADMIN");
+
+    private final String roleName;
+
+    UserRole(String roleName) {
+        this.roleName = roleName;
+    }
+    @Override
+    public String toString() {
+        return roleName;
+    }
 }
